+++ NEWS.md
@@ -1,7 +1,37 @@
-# ARDL 0.1.0.9000
+# ARDL 0.1.1
+
+### New features
+
+* The `multipliers()` function now supports the calculation of short-run and
+interim multipliers.
+
+### Documentation improvements
+
+* Corrected the example about the `search_type` argument in the help file of
+`auto_ardl()`. 
+
+* Minor correction in the help file of `auto_ardl()`.
+
+* Corrected the mathematical formula of the Long-Run multipliers in the help
+file of `multipliers()`.
+
+* Corrected the mathematical formula of the bounds F-test in the help file of
+`bounds_f_test()`.
+
+* Added the mathematical formulas of the short-run and interim multipliers in 
+the help file of `multipliers()`.
+
+* Properly changed the citation info in the `CITATION`, `ardl-package.R` and  
+`DESCRIPTION` files.
+
+* Changed the designer's contact info in the logo.
+
+* Added a small example about short-run multipliers in the `README.md` file.
+
+* Corrected the release date of version 0.1.0 in the `NEWS.md` file.
 
 ---
 
-# ARDL v0.1.0  (09 Apr 2020)
+# ARDL 0.1.0  (10 Apr 2020)
 
 ## Released to CRAN